project/requirements.txt
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cycler==0.10.0
decorator==4.4.2
Flask==1.1.1
idna==2.6
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.17.2
pandas==0.24.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
progressbar2==3.50.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2019.3
requests==2.18.4
scikit-learn==0.19.2
scipy==1.3.1
seaborn==0.9.0
six==1.14.0
traitlets==4.3.3
urllib3==1.22
wcwidth==0.1.8
Werkzeug==1.0.0
wget==3.2